requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
python-dotenv
langchain-text-splitters